updated so that ONLY build-script command is affected by the debian flavor change, and respects environment variable configuration
