+++ src/BuildScriptGeneratorCli/Commands/BuildScriptCommand.cs
@@ -5,11 +5,14 @@
 
 using System;
 using System.IO;
+using System.Linq;
 using McMaster.Extensions.CommandLineUtils;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Microsoft.Oryx.BuildScriptGenerator;
 using Microsoft.Oryx.BuildScriptGenerator.Common;
+using Microsoft.Oryx.BuildScriptGenerator.Exceptions;
 using Microsoft.Oryx.Common.Extensions;
 
 namespace Microsoft.Oryx.BuildScriptGeneratorCli
@@ -95,6 +98,49 @@ namespace Microsoft.Oryx.BuildScriptGeneratorCli
                 appType: this.AppType,
                 scriptOnly: true,
                 properties: this.Properties);
+        }
+
+        internal override IServiceProvider TryGetServiceProvider(IConsole console)
+        {
+            // Don't use the IConsole instance in this method -- override this method in the command
+            // and pass IConsole through to ServiceProviderBuilder to write to the output.
+            var serviceProviderBuilder = new ServiceProviderBuilder(this.LogFilePath)
+                .ConfigureServices(services =>
+                {
+                    var configuration = new ConfigurationBuilder()
+                        .AddEnvironmentVariables()
+                        .Build();
+
+                    services.AddSingleton<IConfiguration>(configuration);
+                })
+                .ConfigureScriptGenerationOptions(opts =>
+                {
+                    this.ConfigureBuildScriptGeneratorOptions(opts);
+
+                    // For debian flavor, we first check for existance of an environment variable
+                    // which contains the os type. If this does not exist, parse the
+                    // FilePaths.OsTypeFileName file for the correct flavor
+                    if (string.IsNullOrWhiteSpace(opts.DebianFlavor))
+                    {
+                        var ostypeFilePath = Path.Join("/opt", "oryx", FilePaths.OsTypeFileName);
+                        if (File.Exists(ostypeFilePath))
+                        {
+                            // these file contents are in the format <OS_type>|<Os_version>, e.g. DEBIAN|BULLSEYE
+                            // we want the Os_version part only, as all lowercase
+                            var fullOsTypeFileContents = File.ReadAllText(ostypeFilePath);
+                            opts.DebianFlavor = fullOsTypeFileContents.Split("|").TakeLast(1).SingleOrDefault().Trim().ToLowerInvariant();
+                        }
+                        else
+                        {
+                            // If we cannot resolve the debian flavor, error out as we will not be able to determine
+                            // the correct SDKs to pull
+                            var errorMessage = $"Error: Image debian flavor not found in DEBIAN_FLAVOR environment variable or the " +
+                                $"{Path.Join("/opt", "oryx", FilePaths.OsTypeFileName)} file. Exiting...";
+                            throw new InvalidUsageException(errorMessage);
+                        }
+                    }
+                });
+            return serviceProviderBuilder.Build();
         }
     }
 }
