+++ src/BuildScriptGeneratorCli/Commands/BuildScriptGeneratorOptionsHelper.cs
@@ -83,29 +83,6 @@ namespace Microsoft.Oryx.BuildScriptGeneratorCli
                 options.ScriptOnly = scriptOnly.Value;
             }
 
-            // For debian flavor, we first check for existance of an environment variable
-            // which contains the os type. If this does not exist, parse the
-            // FilePaths.OsTypeFileName file for the correct flavor
-            if (string.IsNullOrWhiteSpace(options.DebianFlavor))
-            {
-                var ostypeFilePath = Path.Join("/opt", "oryx", FilePaths.OsTypeFileName);
-                if (File.Exists(ostypeFilePath))
-                {
-                    // these file contents are in the format <OS_type>|<Os_version>, e.g. DEBIAN|BULLSEYE
-                    // we want the Os_version part only, as all lowercase
-                    var fullOsTypeFileContents = File.ReadAllText(ostypeFilePath);
-                    options.DebianFlavor = fullOsTypeFileContents.Split("|").TakeLast(1).SingleOrDefault().Trim().ToLowerInvariant();
-                }
-                else
-                {
-                    // If we cannot resolve the debian flavor, error out as we will not be able to determine
-                    // the correct SDKs to pull
-                    var errorMessage = $"Error: Image debian flavor not found in DEBIAN_FLAVOR environment variable or the " +
-                        $"{Path.Join("/opt", "oryx", FilePaths.OsTypeFileName)} file. Exiting...";
-                    throw new InvalidUsageException(errorMessage);
-                }
-            }
-
             options.Properties = ProcessProperties(properties);
         }
